Wed, Dec 27, 2023, 5:23 AM +08:00

+++ hours/hours.c
@@ -44,7 +44,7 @@ float calc_hours(int hours[], int weeks, char output)
         {
             total = total + hours[i];
         }
-        return total / weeks;
+        return total / (float) weeks;
     }
     return 0;
 }
